TextPic.cs was not working for href.
Unity-UI-Extensions helped me a lot thanks for creating it.

Today i was trying href tag provided with TextPic.cs, but unfortunately it was not working.

I figured out that, OnPopulateMesh() was creating hrefInfo.boxes but at the end when OnPopulateMesh() calls UpdateQuadImage(), hrefInfo.boxes.count was coming 0. This was little strange but i saw that GetOutputText() was re-creating m_HrefInfos array from scratch but it didn't include the code to add boxes, so this was the issue.

so what i did is, i commented this line m_HrefInfos.Clear(); in GetOutputText() and added few lines of code to check if m_HrefInfos elements exists than don't destroy it, just update it with new values.
This helped in resolving the issue. Now i was getting onClick event on href tag.

Thanks

--HG--
branch : sushanta1991/textpiccs-was-not-working-for-href-unit-1474979920532