Updated TextPic.   Recent Unity updates seem to have changed the Event System behaviour and all click events are broken when a TextPic is placed as the child of a button. (gave up figuring out why)
Changed implementation, so that when TextPic is added as a child of a button, it adds a CanvasGroup and allows clicks to pass through it.  Downside is that hyperlinks don't work. TextPic on buttons is purely decorative.

The Button issue disables the ability to highlight links, so updated implementation to allow a "Selectable" component to be added to the TextPic to set highlight colours.

Resolves #110