Merged in development (pull request #122)
Development

* Merged in updated_new_input_system_check (pull request #117)

Updated the new Input system switch and tested against 2021

* Updated LineRenderer to initialise with a 1 point array, this resolves an issue with Unity putting the previous controls vertext array in an uninitialised control.
Also updated examples checkout

* Merged in UpdatedLineRender (pull request #118)

UILineRender now initilises with a point of 1 to correct Unity issue

* Applied J.R. Mitchell's fix for the Accordion Controls/Accordion/AccordionElement.cs

resolves: #364

* Merged in AccordionFIx (pull request #119)

Update to Accordion to allow for prefab Accordion Elements

* Resolved issue where the Content Scroll snap would cause issue with only 1 child.
Also tested and fixed working with no children.

Resolves #362

* Merged in content_scroll_snap_fix (pull request #120)

ContentScrollSnap update

* Updated the PaginationManager to override if the ScrollSnap is in motion.
BuyerBeware, Unity's toggles can get in to a state where they are clicked but not clicked.

* Merged in PaginationFix (pull request #121)

Updated the PaginationManager to override if the ScrollSnap is in motion.

* Updated release notes and bumped version to release